+++ frontend/src/pages/knowledgeGraph/PageCite.vue
@@ -94,7 +94,7 @@
 <script setup lang="ts">
 import AppBreadcrumb from "@/components/AppBreadcrumb.vue";
 import AppCitation from "@/components/AppCitation.vue";
-import PageTitle from "@/components/PageTitle.vue";
+import PageTitle from "@/components/ThePageTitle.vue";
 </script>
 <style lang="scss" scoped>
 $wrap: 1000px;
